backend/services/document-processor/requirements.txt
google-cloud-aiplatform
google-cloud-documentai
functions-framework
chromadb
gcsfs
functions-framework
google-cloud-storage
langchain-google-vertexai
langchain_text_splitters